backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
numpy==2.2.5
pillow==11.2.1
pydicom==3.0.1
PyJWT==2.9.0
sqlparse==0.5.3
